requirements.txt
# LabelWatch AI - FDA DailyMed scraper & labeling tracker
# Day 1-2 MVP: RSS + watchlist + impact report output

requests>=2.31.0
feedparser>=6.0.10
beautifulsoup4>=4.12.0
lxml>=5.0.0
python-dotenv>=1.0.0
streamlit>=1.28.0
fpdf2>=2.7.0
